+++ apps/ui-dev/component/navigations/tabs/LabelView.jsx
@@ -2,14 +2,14 @@ import { Grid, Tab, TabItem, VBox } from "@elf-framework/ui";
 
 export function LabelView() {
   return (
-    <Grid columns={["50%", 1]} gap={30}>
+    <Grid columns={1} gap={30}>
       <VBox
         style={{
           gap: 30,
           backgroundColor: "var(--color-gray-0)",
           padding: "30px 20px",
           fontSize: 13,
-          height: 100,
+          height: 150,
         }}
       >
         <div