+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/ui
 
+## 0.0.19
+
+### Patch Changes
+
+- add withLabel option in switch
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.19
+
 ## 0.0.13
 
 ### Patch Changes